# internal/db/handler/indexer/list_indexers.go
package indexer

import (
	"gorgon/internal/db/repository"
	"gorgon/pkg/schemas"

	"github.com/labstack/echo/v4"
)

// @BasePath /api/v1

// @Summary List Indexers
// @Description List all added indexers
// @Tags Database/Indexer
// @Produce json
// @Success 200 {object} schemas.DefaultResponse
// @Failure 400 {object} schemas.ErrorResponse
// @Failure 500 {object} schemas.ErrorResponse
// @Router /database/indexer [get]
func ListIndexers(c echo.Context) error {
	indexerRepo := repository.NewIndexerRepository()

	indexers, err := indexerRepo.List()
	if err != nil {
		return err
	}

	schemas.SendSucess(c, "List Indexers", indexers)
	return nil
}